Move remaining util functions to tasks module
https://pagure.io/freeipa/issue/6798
