+++ ipatests/test_integration/test_trust.py
@@ -22,7 +22,6 @@ import re
 
 from ipatests.test_integration.base import IntegrationTest
 from ipatests.pytest_plugins.integration import tasks
-from ipatests.test_integration import util
 from ipaplatform.paths import paths
 
 
@@ -189,8 +188,8 @@ class TestBasicADTrust(ADTrustBase):
                                 stdin_text=original_passwd)
 
         # change password for the user to be able to kinit
-        util.ldappasswd_user_change(ipauser, original_passwd, new_passwd,
-                                    self.master)
+        tasks.ldappasswd_user_change(ipauser, original_passwd, new_passwd,
+                                     self.master)
 
         # try to kinit as ipauser
         self.master.run_command(
